Abandoning JSON configuration for commands and deals.
Decided to go with an embedded, local database for both copypasta commands and deals. Will be using H2 as the database solution in this project, after going through around a day of thinking.

Due to this change, there is no longer a need for the JSON configuration files for commands. Will be working on implementing objects to handle the data used and worked with for both commands and deals at the same time, before implementing the needed database work, letting the program interact with our database.